refactor(classification): split symbols module and add MSVC idempotency test
Addresses two non-blocking review follow-ups from PR #177:

- symbols.rs exceeded the 500-line soft limit (674 lines, mostly test
  growth). Split into symbols/mod.rs (production, 306 lines) and
  symbols/tests.rs (390 lines), mirroring the existing patterns/ module
  directory convention in classification/.
- Add test_demangle_msvc_idempotent covering the MSVC double-demangle
  no-op, matching the existing Rust-path test_demangle_idempotent. The
  no-op was mechanically guaranteed by the is_mangled() early-exit but
  not directly tested.

No behavior change. fmt + clippy clean; all symbol tests pass.
